+++ examples/seismic/tti/operators.py
@@ -368,10 +368,10 @@ def ForwardOperator(model, source, receiver, time_order=2, space_order=4,
                                            model.delta, model.theta, model.phi)
 
     # Create symbols for forward wavefield, source and receivers
-    u = TimeData(name='u', grid=model.grid, dtype=model.dtype,
+    u = TimeData(name='u', grid=model.grid,
                  save=save, time_dim=source.nt if save else None,
                  time_order=time_order, space_order=space_order)
-    v = TimeData(name='v', grid=model.grid, dtype=model.dtype,
+    v = TimeData(name='v', grid=model.grid,
                  save=save, time_dim=source.nt if save else None,
                  time_order=time_order, space_order=space_order)
     src = PointSource(name='src', ntime=source.nt, ndim=source.ndim,
